requirements.txt
pyyaml
google-analytics-data>=0.18.0
pandas>=2.0.0
matplotlib>=3.7.0
scipy>=1.10.0
